Se comparte con un código por query param

+++ src/utils/useMaterias86.js
@@ -1,13 +1,38 @@
 import { useState } from 'react';
+import plan86 from '../plan_86.json';
+
+function decode(code) {
+  let bits = "";
+  for (let i = 0; i < code.length; i++) {
+    bits += parseInt(code[i], 16).toString(2).padStart(4, 0);
+  }
+
+  let materias = [];
+  materias = materias.concat(plan86.obligatorias.filter((materia, idx) => bits[idx] === "1"));
+  let offset = plan86.obligatorias.length;
+  plan86.orientaciones.forEach(orientacion => {
+    materias = materias.concat(orientacion.materias.filter((materia, idx) => bits[idx+offset] === "1"));
+    offset += orientacion.materias.length;
+  });
+  materias = materias.concat(plan86.electivas.filter((materia, idx) => bits[idx+offset] === "1"));
+  return materias;
+}
 
 export function useMaterias86(key, initialValue) {
     // State to store our value
     // Pass initial state function to useState so logic is only executed once
+    const [readOnly, setReadOnly] = useState(false);
     const [storedValue, setStoredValue] = useState(() => {
       if (typeof window === "undefined") {
         return initialValue;
       }
       try {
+        const shareCode = new URLSearchParams(document.location.search).get("code");
+        if (shareCode !== null) {
+          const values = decode(shareCode);
+          setReadOnly(true);
+          return values;
+        }
         // Get from local storage by key
         const item = window.localStorage.getItem(key);
         // Parse stored json or if none return initialValue
@@ -27,6 +52,10 @@ export function useMaterias86(key, initialValue) {
           value instanceof Function ? value(storedValue) : value;
         // Save state
         setStoredValue(valueToStore);
+
+        if (readOnly)
+          return;
+
         // Save to local storage
         if (typeof window !== "undefined") {
           window.localStorage.setItem(key, JSON.stringify(valueToStore));
@@ -36,5 +65,5 @@ export function useMaterias86(key, initialValue) {
         console.log(error);
       }
     };
-    return [storedValue ??[], setValue];
+    return [storedValue ??[], setValue, readOnly];
   }
